feat: add authenticated routing and api client
