+++ src/App.css
@@ -84,6 +84,25 @@ main.app {
     transform var(--transition-fast);
 }
 
+.app__logout-button {
+  justify-self: end;
+  padding: 0.55rem 1rem;
+  border-radius: 999px;
+  border: 1px solid rgba(255, 255, 255, 0.22);
+  background: rgba(255, 255, 255, 0.06);
+  color: var(--color-text-primary);
+  font-weight: 600;
+  cursor: pointer;
+  transition: background-color var(--transition-fast), border-color var(--transition-fast), color var(--transition-fast);
+}
+
+.app__logout-button:hover,
+.app__logout-button:focus-visible {
+  background: rgba(255, 255, 255, 0.12);
+  border-color: rgba(255, 255, 255, 0.3);
+  color: var(--color-text-secondary);
+}
+
 .app__nav-link::after {
   content: '';
   position: absolute;
@@ -215,7 +234,7 @@ main.app {
     grid-template-areas:
       'logo toggle'
       'nav nav'
-      'cta cta';
+      'logout logout';
     align-items: center;
     justify-items: stretch;
     row-gap: var(--space-2);
@@ -262,6 +281,12 @@ main.app {
     width: 100%;
     justify-content: flex-start;
     padding: 0.6rem 0.75rem;
+  }
+
+  .app__logout-button {
+    grid-area: logout;
+    width: 100%;
+    padding: 0.7rem 1rem;
   }
 }
