+++ src/App.jsx
@@ -1,233 +1,18 @@
-import { useCallback, useEffect, useState } from 'react';
+import { BrowserRouter, Navigate, Route, Routes } from 'react-router-dom';
+import AuthPage from './features/auth/AuthPage.jsx';
+import HomePage from './features/home/HomePage.jsx';
+import PrivateRoute from './components/PrivateRoute.js';
 
-import Section from './components/Section.jsx';
-import Footer from './components/Footer.jsx';
-import Hero from './features/Hero/Hero.jsx';
-import Overview from './features/Overview/Overview.jsx';
-import RegistrationCta from './features/RegistrationCta/RegistrationCta.jsx';
-import Sponsors from './features/Sponsors/Sponsors.jsx';
-import TeamShowcase from './features/team-showcase/TeamShowcase.js';
-import SocialLinks from './features/SocialLinks/SocialLinks.jsx';
-import heroConfig from './features/Hero/config.json';
-import overviewConfig from './features/Overview/config.json';
-import UpcomingMatches from './features/UpcomingMatches/UpcomingMatches.jsx';
-import upcomingMatchesConfig from './features/UpcomingMatches/config.json';
-import MatchResults from './features/MatchResults/MatchResults.jsx';
-import matchResultsConfig from './features/MatchResults/config.json';
-import registrationCtaConfig from './features/RegistrationCta/config.json';
-import sponsorsConfig from './features/Sponsors/config.json';
-import socialLinksConfig from './features/SocialLinks/config.json';
-import QualificationsTable from './features/QualificationsTable/QualificationsTable.jsx';
-import qualificationsConfig from './features/QualificationsTable/config.json';
-import logoImage from './ChatGPT Image 20 окт. 2025 г., 22_02_10.png';
-import './App.css';
-
-const App = () => {
-  const handleSponsorFormSubmit = useCallback(async (formData) => {
-    // Здесь мог бы быть вызов API или интеграция с сервисом отправки.
-    console.log('Sponsor form submitted', formData);
-  }, []);
-
-  const sections = [
-    {
-      id: 'hero',
-      component: <Hero data={heroConfig} disableVideoOnMobile={heroConfig.media?.disableOnMobile} />,
-      variant: 'hero',
-      hideTitle: true,
-      fullBleed: true,
-      navLabel: 'Главная',
-    },
-    {
-      id: 'overview',
-      title: 'Обзор YarCyberSeason',
-      component: <Overview data={overviewConfig} />,
-      navLabel: 'Обзор',
-      variant: 'overview',
-      background: (
-        <div className="overview-background">
-          <div className="overview-background__blur overview-background__blur--primary" />
-          <div className="overview-background__blur overview-background__blur--secondary" />
-          <div className="overview-background__grain" />
-        </div>
-      ),
-    },
-    {
-      id: 'team-showcase',
-      title: 'Команды сезона',
-      component: <TeamShowcase />,
-      navLabel: 'Команды',
-      variant: 'team-showcase',
-      hideTitle: true,
-    },
-    {
-      id: 'upcoming-matches',
-      component: <UpcomingMatches data={upcomingMatchesConfig} />,
-      navLabel: 'Матчи',
-      variant: 'upcoming-matches',
-      hideTitle: true,
-    },
-    {
-      id: 'qualifications',
-      component: (
-        <QualificationsTable data={qualificationsConfig} matchResults={matchResultsConfig} />
-      ),
-      navLabel: 'Таблица',
-      variant: 'qualifications-table',
-      hideTitle: true,
-      fullBleed: true,
-    },
-    {
-      id: 'match-results',
-      component: <MatchResults data={matchResultsConfig} />,
-      navLabel: 'Результаты',
-      variant: 'match-results',
-      hideTitle: true,
-    },
-    {
-      id: 'registration',
-      component: <RegistrationCta data={registrationCtaConfig} />,
-      navLabel: 'Регистрация',
-      variant: 'registration-cta',
-      hideTitle: true,
-    },
-    {
-      id: 'social',
-      title: 'Социальные каналы',
-      component: <SocialLinks data={socialLinksConfig} />,
-      navLabel: 'Соцсети',
-      variant: 'social-links',
-      hideTitle: true,
-    },
-    {
-      id: 'sponsors',
-      title: 'Партнёры и спонсоры',
-      component: (
-        <Sponsors
-          data={sponsorsConfig}
-          onSponsorFormSubmit={handleSponsorFormSubmit}
-        />
-      ),
-      navLabel: 'Партнёры',
-      variant: 'sponsors',
-    },
-  ];
-
-  const navigationItems = sections
-    .filter((section) => Boolean(section.navLabel))
-    .map((section) => ({
-      id: section.id,
-      label: section.navLabel,
-    }));
-
-  const THEME_STORAGE_KEY = 'ycs-theme';
-
-  const getInitialTheme = useCallback(() => {
-    if (typeof window === 'undefined') {
-      return 'default';
-    }
-
-    return window.localStorage.getItem(THEME_STORAGE_KEY) || 'default';
-  }, []);
-
-  const [theme, setTheme] = useState(getInitialTheme);
-  const [isMenuOpen, setIsMenuOpen] = useState(false);
-
-  useEffect(() => {
-    const handleResize = () => {
-      if (window.innerWidth > 768 && isMenuOpen) {
-        setIsMenuOpen(false);
-      }
-    };
-
-    window.addEventListener('resize', handleResize);
-
-    return () => {
-      window.removeEventListener('resize', handleResize);
-    };
-  }, [isMenuOpen]);
-
-  useEffect(() => {
-    if (typeof document === 'undefined') {
-      return;
-    }
-
-    const root = document.documentElement;
-    root.dataset.theme = theme;
-
-    if (typeof window !== 'undefined') {
-      window.localStorage.setItem(THEME_STORAGE_KEY, theme);
-    }
-  }, [theme]);
-
-  const toggleMenu = () => {
-    setIsMenuOpen((prev) => !prev);
-  };
-
-  const handleNavLinkClick = () => {
-    setIsMenuOpen(false);
-  };
-
-  const handleThemeToggle = () => {
-    setTheme((prevTheme) => (prevTheme === 'feminine' ? 'default' : 'feminine'));
-  };
-
-  const isFeminineTheme = theme === 'feminine';
-
-  return (
-    <main className="app">
-      <header className="app__header">
-        <a className="app__logo" href="#hero" aria-label="Перейти к началу страницы">
-          <img className="app__logo-image" src={logoImage} alt="Логотип YarCyberSeason" />
-          <span className="app__logo-text" aria-hidden="true">ЯрКиберСезон 25/26</span>
-        </a>
-        <button
-          type="button"
-          className={`app__menu-toggle${isMenuOpen ? ' app__menu-toggle--open' : ''}`}
-          aria-label={isMenuOpen ? 'Скрыть меню' : 'Открыть меню'}
-          aria-expanded={isMenuOpen}
-          aria-controls="app-navigation"
-          onClick={toggleMenu}
-        >
-          <span className="app__menu-toggle-line" aria-hidden="true" />
-          <span className="app__menu-toggle-line" aria-hidden="true" />
-          <span className="app__menu-toggle-line" aria-hidden="true" />
-        </button>
-        <nav
-          id="app-navigation"
-          className={`app__nav${isMenuOpen ? ' app__nav--open' : ''}`}
-          aria-label="Навигация по секциям YarCyberSeason"
-        >
-          <ul className="app__nav-list">
-            {navigationItems.map(({ id, label }) => (
-              <li key={id} className="app__nav-item">
-                <a className="app__nav-link" href={`#${id}`} onClick={handleNavLinkClick}>
-                  {label}
-                </a>
-              </li>
-            ))}
-          </ul>
-        </nav>
-      </header>
-      {sections.map((section) => {
-        const { id, title, component, variant, hideTitle, fullBleed, background, modifiers } = section;
-        return (
-          <Section
-            key={id}
-            id={id}
-            title={title}
-            variant={variant}
-            hideTitle={hideTitle}
-            fullBleed={fullBleed}
-            background={background}
-            modifiers={modifiers}
-          >
-            {component}
-          </Section>
-        );
-      })}
-      <Footer isFeminineTheme={isFeminineTheme} onThemeToggle={handleThemeToggle} />
-    </main>
-  );
-};
+const App = () => (
+  <BrowserRouter>
+    <Routes>
+      <Route path="/auth" element={<AuthPage />} />
+      <Route element={<PrivateRoute />}>
+        <Route path="/" element={<HomePage />} />
+      </Route>
+      <Route path="*" element={<Navigate to="/" replace />} />
+    </Routes>
+  </BrowserRouter>
+);
 
 export default App;
